+++ packages/perseus-editor/src/widgets/expression-editor.tsx
@@ -1,5 +1,5 @@
 import * as KAS from "@khanacademy/kas";
-import {components, Expression} from "@khanacademy/perseus";
+import {ApiOptions, components, Expression} from "@khanacademy/perseus";
 import {
     PerseusExpressionAnswerFormConsidered,
     deriveExtraKeys,
@@ -17,10 +17,10 @@ import _ from "underscore";
 
 import styles from "./expression-editor.module.css";
 
+import type {APIOptions} from "@khanacademy/perseus";
 import type {
     PerseusExpressionWidgetOptions,
     LegacyButtonSets,
-    ExpressionDefaultWidgetOptions,
     PerseusExpressionAnswerForm,
     PerseusExpressionUserInput,
 } from "@khanacademy/perseus-core";
@@ -31,6 +31,7 @@ const {ButtonGroup, InfoTip} = components;
 type Props = {
     widgetId?: string;
     value?: string;
+    apiOptions: APIOptions;
     onChange: (newValues: Partial<PerseusExpressionWidgetOptions>) => void;
 } & Omit<PerseusExpressionWidgetOptions, "buttonsVisible">;
 
@@ -61,8 +62,10 @@ interface State {
 class ExpressionEditor extends React.Component<Props, State> {
     static widgetName = "expression" as const;
 
-    static defaultProps: ExpressionDefaultWidgetOptions =
-        expressionLogic.defaultWidgetOptions;
+    static defaultProps = {
+        ...expressionLogic.defaultWidgetOptions,
+        apiOptions: ApiOptions.defaults,
+    };
 
     constructor(props: Props) {
         super(props);
@@ -285,6 +288,7 @@ class ExpressionEditor extends React.Component<Props, State> {
     };
 
     render(): React.ReactNode {
+        const editingDisabled = this.props.apiOptions.editingDisabled ?? false;
         const answerOptions: React.JSX.Element[] = this.props.answerForms.map(
             (ans: AnswerForm, index: number) => {
                 const expressionProps: Partial<
@@ -312,6 +316,7 @@ class ExpressionEditor extends React.Component<Props, State> {
                     <AnswerOption
                         key={ans.key}
                         considered={ans.considered}
+                        editingDisabled={editingDisabled}
                         // eslint-disable-next-line no-restricted-syntax
                         expressionProps={expressionProps as any}
                         form={ans.form}
@@ -457,7 +462,11 @@ class ExpressionEditor extends React.Component<Props, State> {
 
                 <View className={styles.answersGroup}>
                     <View className={styles.answersList}>{answerOptions}</View>
-                    <Button size="small" onClick={this.newAnswer}>
+                    <Button
+                        size="small"
+                        disabled={editingDisabled}
+                        onClick={this.newAnswer}
+                    >
                         Add new answer
                     </Button>
                 </View>
@@ -475,6 +484,7 @@ const findNextIn = function <T>(arr: ReadonlyArray<T>, val: T) {
 
 interface AnswerOptionProps {
     considered: (typeof PerseusExpressionAnswerFormConsidered)[number];
+    editingDisabled: boolean;
     expressionProps: React.ComponentProps<typeof Expression>;
 
     // Must the answer have the same form as this answer.
@@ -522,10 +532,12 @@ class AnswerOption extends React.Component<
     };
 
     render(): React.ReactNode {
+        const {editingDisabled} = this.props;
         const removeButton = this.state.deleteFocused ? (
             <>
                 <Button
                     size="small"
+                    disabled={editingDisabled}
                     onClick={this.handleImSure}
                     actionType="destructive"
                 >
@@ -533,6 +545,7 @@ class AnswerOption extends React.Component<
                 </Button>
                 <Button
                     size="small"
+                    disabled={editingDisabled}
                     onClick={this.handleCancelDelete}
                     kind="secondary"
                 >
@@ -542,6 +555,7 @@ class AnswerOption extends React.Component<
         ) : (
             <Button
                 size="small"
+                disabled={editingDisabled}
                 onClick={this.handleDelete}
                 actionType="destructive"
                 kind="tertiary"
@@ -556,6 +570,7 @@ class AnswerOption extends React.Component<
                 <ButtonGroup
                     onChange={this.toggleConsidered}
                     allowEmpty={false}
+                    disabled={editingDisabled}
                     value={this.props.considered}
                     selectedButtonStyle={
                         consideredButtonStyles[this.props.considered]