+++ packages/perseus/src/__tests__/renderer.test.tsx
@@ -1,3 +1,4 @@
+// TODO(LEMS-4304): feature flag cleanup - remove this file.
 import {describe, beforeAll, beforeEach, afterEach, it} from "@jest/globals";
 import {
     generateTestPerseusItem,